feature: ProductDetail 제거하고 ProductQueryHandler로 병합
