+++ src/main/java/com/github/thundermarket/thundermarket/controller/ProductController.java
@@ -3,7 +3,6 @@ package com.github.thundermarket.thundermarket.controller;
 import com.github.thundermarket.thundermarket.aspect.SessionUserParam;
 import com.github.thundermarket.thundermarket.domain.*;
 import com.github.thundermarket.thundermarket.service.ProductCommandHandler;
-import com.github.thundermarket.thundermarket.service.ProductDetailService;
 import com.github.thundermarket.thundermarket.service.ProductQueryHandler;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,12 +15,10 @@ public class ProductController {
 
     private final ProductQueryHandler productQueryHandler;
     private final ProductCommandHandler productCommandHandler;
-    private final ProductDetailService productDetailService;
 
-    public ProductController(ProductQueryHandler productQueryHandler, ProductCommandHandler productCommandHandler, ProductDetailService productDetailService) {
+    public ProductController(ProductQueryHandler productQueryHandler, ProductCommandHandler productCommandHandler) {
         this.productQueryHandler = productQueryHandler;
         this.productCommandHandler = productCommandHandler;
-        this.productDetailService = productDetailService;
     }
     @GetMapping("/api/v1/products")
     public ResponseEntity<ProductsResponse> products(
@@ -33,7 +30,7 @@ public class ProductController {
 
     @GetMapping("/api/v1/products/{id}")
     public ResponseEntity<ProductDetailResponse> productDetail(@PathVariable("id") Long id) {
-        return new ResponseEntity<>(productDetailService.productDetail(id), HttpStatus.OK);
+        return new ResponseEntity<>(productQueryHandler.productDetail(id), HttpStatus.OK);
     }
 
     @PostMapping("/api/v1/products")
